Add in comment that this is oc_legacy_relay
Note that if this relay is used instead of oc_relay it should still be named
oc_relay. We might have to make seperate installers for the two relays
so people can switch. I prefer this one as it seems to work better for
badly coded RLV objects (doesn't discard permission).