Let oc_settings use linksetdata #101
Includes a 'Dump LSD' command in the Settings dialog, to show what's
currently stored. You can compare to the 'Print' to see if everything
is correct.

On 'Load Card', linksetdata will be cleared before loading, at least for
now. This is maybe not desirable in the final release, otherwise one can
purge all owners too easily.